Update __init__.py
The get_template() function used .find() method to find the dot that separates the filename from the extension.

Because of that, headers didn't work with files that contained a dot in the filename, i.e.:

`Robot.class.cpp` doesn't match the 'cpp' extension, because .find() will return the index of the first dot. The function then believes the extension is 'class.cpp'. It is, therefore, impossible to have a working header in this kind of file.

If we use .rfind() instead, these files will be matched correctly.